Add snippet-assisted hover
The `snippets/gsl.json` file contains a wealth of information on
commands. This commit allows the hover provider to match snippets
based on the first word in their prefix, displaying the corresponding
snippet descriptions in markdown.

In the case where multiple snippets match, all of the descriptions
are shown. One such case is `callmatch`. A more sophisticated
implementation of this commit might resolve to a single snippet, but
it may also be desirable to show the user alternative usages. I'm on the
fence.

It may also be cleaner to separate the descriptions from the snippets,
but that is a much larger diff, and would require dynamically building
the snippets from multiple files. If that direction is desired, then it
might be worth considering moving `gsl.json` to a typescript file.

Finally, note that this commit adds `resolveJsonModule: true` to the
tsconfig so that json files are importable in typescript.