Add safety guard to script uploads
This addresses an issue where a user could inadvertantly overwrite
a remote script without recognizing that they are either:
A) Working with an outdated copy of the script, or
B) Not the last modifier of the script

(note that the code for A existed in the past but was removed
 in commit d2934fb967 - rewrite js to ts)

The logic for this commit's safety guard is similar to the code that
previously existed. When a script is downloaded, the properties are
stored in `context.globalState`. These properties include the last
modified date, the script author, as well as other things. When a
script is subsequently uploaded the script properties are again read
from the server, and a confirmation is required if:
A) There is no local last modified date. Possibly the script was
   download prior to this guard's implementation, or the script
   was downloaded manually out-of-band.
B) The server's last modified date does not match the local
   last modified date.
C) The server's last modifying account does not match the local
   last modifying account.

Finally, I removed the start-of-line anchor from the `rx_modified`
regex used for capturing the last modified time. This is because when
`/mv` is used instead of `/ms`, the modified time line is captured
as follows: `GN>On Mon Apr  9 16:11:32 2018`.