Fix unwanted file navigations
This fixes an issue where users would accidentally navigate to
another file if a matchmarker definition is requested that is
not available locally. The previous code would download the
script and navigate to the target matchmarker, but this is often
not desired when a definition is requested. Note that users can
still right click on the matchmarker and "go to definition", as
the definition is now provided asynchronously in the case where
the script needs to be downloaded.