Restore "Compile OK" status bar message
This message was inadvertantly removed by 92ab1d, which added a call
to `getScriptProperties()` in `uploadScript()`. As a result, the
`getScriptProperties()` call would overwrite the "Compile OK" message
with a "Script xxx was last modified..." message.

This commit pulls instances of `window.setStatusBarMessage()` out of
`getScriptProperties()` and `uploadScript()` and into the top-level
command methods `commandCheckDate()` and `commandUploadScript()`. This
allows inner methods to call one another without worrying about sending
a message to the user (note that error messages will still be surfaced -
it may be a good idea to pull that logic out as well in the future).

This commit also displays bytes remaining against the script size
limit.