Record script properties using script number
This solves an issue where the extension wouldn't distinguish between
a script downloaded with it's verb name versus it's script number.
Backwards compatibility with the old globalState values should be
preserved.

Rather than storing all script properties, we now only store the
last modifier and last modified time. This is because those properties
are the only ones that we use, and because the output of `/ss` does
not give us all properties - we want to avoid a roundtrip with
`/m[sv]`.

This commit also increases the type safety of `editorClient.ts`:
- Distinguish between the return type of `checkScript` and
  `modifyScript`. The former is a subset of the latter.
- Add some type safety to those same methods while building up
  the return value. Two `as any` casts still remain but these would
  need a larger change to address, likely requiring one regex per
  property.
- Properly mark `ScriptProperties.new` as optional