Update "instances out of sync" message
Soften from warning message to informational message. Clarify that
this is being shown because instances are out of sync. Finally,
only show this message when the last modifier is not the current user.

This is in response to feedback that "Status: newer script in dev"
message was confusing and concerning. A user thought the message
meant that the script they were downloading was somehow older
than what was in development.