Add script-out-of-date button
This button appears on the status bar and gives the user information
about whether the current script is out of date relative to the
development server. If the script is out of date, then clicking the
button will allow the user to update the script, view a file comparison
between the local script and the script in development, and to stop
monitoring the script until a new version is seen. If the user has
previously marked the script as ignored, clicking the button will
allow the user to resume monitoring.

In the case where the contents of a file matches, the local modification
time will be updated to the server modification time. This is VERY
IMPORTANT, because it prevents the file from being shown as out of date
in the case where the user proceeds to modify the local file.

Fixes #15