Add "frozen script" warning
The `/FREEZE` command allows for marking scripts as frozen, i.e.
not eligible for deploy/rollin. This commit surfaces a warning in
the status bar for those scripts. This warning should help shorten
the time to identification of frozen scripts, hopefully prior to
changes being made.

Because the list of frozen scripts is short and relatively static,
we use periodic polling to maintain a complete list. This is in
contrast to an approach like the out-of-date button, which checks
script mod times each time the active text editor changes.

This commit also adds `EditorClient.executeCommand`, which is a
generic mechanism for sending commands to the game and capturing
corresponding outputs.