Add `GSLExtension.matchesRemoteAccount`
Centralizes the logic for comparing the local account name with the
remote account name. This uses `startsWith` because the remote
account names are truncated. This commit also updates updates the
download script path to use this new method, which fixes a bug where
"instances out of sync" wouldn't be shown for users with accounts
longer than 12 characters.