Avoid unnecessary calls to `/ss checkedit`
`/ss checkedit` is used when downloading scripts to check if instances
are out of sync and display this information to the user. However, the
informational message is only shown if the last modifier matches the
current user. Therefore this commit avoids `/ss checkedit` if the last
modifier is not the current user.