Fix reconnect issue
Yusri and Vanah reported an issue where the extension would not
reconnect once severed. I finally tracked it down to a `not connected`
message. This message is shown when `cleanupServer` has been called
in `BaseGameClient`, setting `this.server` to `undefined`.

This commit adds a check to see if the server connection exists and
restores it if not.