Rework Align Comment Code Action
- Drop column 90 as a target column
- Instead always aim for column 64
- In the case where comments are large enough that column 64 is not
  achievable, aim for as high a column as possible.

The net effect of these changes is to aim for as close to column 64
as possible. This fixes a bug where lines with content length 62
would receive comment column 90, and it also behaves closer to my
subjective expectation that comments should align as close together
as possible, e.g. in this case:
```
  msgp "foo"                                                   ! aligned at 64
  msgp "bar"                                                   ! also aligned at 64
  callmatch must_match "$APPLY_ARDOR_OF_THE_SCOURGE_NP0" in 24310                        ! aligned at 90
  msgp "biz"                                                   ! this too is aligned at 64
```

Here we would much prefer that the comment on line 3 is like so:
```
  callmatch must_match "$APPLY_ARDOR_OF_THE_SCOURGE_NP0" in 24310 ! aligned at 90
```